gh-1573 - Repackaged Extract functions, added constructor
