+++ core/data/src/test/java/uk/gov/gchq/gaffer/data/graph/function/walk/ExtractWalkEdgesFromHopTest.java
@@ -13,53 +13,46 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.gchq.gaffer.data.graph.function;
+package uk.gov.gchq.gaffer.data.graph.function.walk;
 
-import com.google.common.collect.Sets;
 import org.junit.Test;
 
 import uk.gov.gchq.gaffer.commonutil.TestGroups;
 import uk.gov.gchq.gaffer.data.element.Edge;
 import uk.gov.gchq.gaffer.data.element.Entity;
 import uk.gov.gchq.gaffer.data.graph.Walk;
+import uk.gov.gchq.gaffer.data.graph.function.walk.ExtractWalkEdgesFromHop;
 
 import java.util.Set;
 import java.util.function.Function;
 
-import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.contains;
 import static org.junit.Assert.assertThat;
 
-public class ExtractWalkEntitiesTest {
+public class ExtractWalkEdgesFromHopTest {
     private static final Edge EDGE_AB = new Edge.Builder().group(TestGroups.EDGE).source("A").dest("B").directed(true).build();
     private static final Edge EDGE_BC = new Edge.Builder().group(TestGroups.EDGE).source("B").dest("C").directed(true).build();
     private static final Edge EDGE_CA = new Edge.Builder().group(TestGroups.EDGE).source("C").dest("A").directed(true).build();
 
-    private static final Entity ENTITY_A = new Entity.Builder().group(TestGroups.ENTITY).vertex("A").build();
     private static final Entity ENTITY_B = new Entity.Builder().group(TestGroups.ENTITY).vertex("B").build();
     private static final Entity ENTITY_C = new Entity.Builder().group(TestGroups.ENTITY).vertex("C").build();
 
     @Test
-    public void shouldReturnEntitiesFromWalkObject() {
+    public void shouldReturnEdgesFromWalkObject() {
         // Given
-        final Function<Walk, Iterable<Set<Entity>>> function = new ExtractWalkEntities();
+        final Function<Walk, Set<Edge>> function = new ExtractWalkEdgesFromHop(1);
         final Walk walk = new Walk.Builder()
-                .entity(ENTITY_A)
                 .edge(EDGE_AB)
                 .entity(ENTITY_B)
                 .edge(EDGE_BC)
                 .entity(ENTITY_C)
                 .edge(EDGE_CA)
-                .entity(ENTITY_A)
                 .build();
 
         // When
-        final Iterable<Set<Entity>> results = function.apply(walk);
+        final Set<Edge> results = function.apply(walk);
 
         // Then
-        assertThat(results, containsInAnyOrder(
-                Sets.newHashSet(ENTITY_A),
-                Sets.newHashSet(ENTITY_B),
-                Sets.newHashSet(ENTITY_C),
-                Sets.newHashSet(ENTITY_A)));
+        assertThat(results, contains(EDGE_BC));
     }
 }
