+++ core/data/src/test/java/uk/gov/gchq/gaffer/data/graph/function/walk/ExtractWalkEntitiesFromHopTest.java
@@ -13,23 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.gchq.gaffer.data.graph.function;
+package uk.gov.gchq.gaffer.data.graph.function.walk;
 
-import com.google.common.collect.Sets;
 import org.junit.Test;
 
 import uk.gov.gchq.gaffer.commonutil.TestGroups;
 import uk.gov.gchq.gaffer.data.element.Edge;
 import uk.gov.gchq.gaffer.data.element.Entity;
 import uk.gov.gchq.gaffer.data.graph.Walk;
+import uk.gov.gchq.gaffer.data.graph.function.walk.ExtractWalkEntitiesFromHop;
 
 import java.util.Set;
 import java.util.function.Function;
 
-import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.contains;
 import static org.junit.Assert.assertThat;
 
-public class ExtractWalkEdgesTest {
+public class ExtractWalkEntitiesFromHopTest {
     private static final Edge EDGE_AB = new Edge.Builder().group(TestGroups.EDGE).source("A").dest("B").directed(true).build();
     private static final Edge EDGE_BC = new Edge.Builder().group(TestGroups.EDGE).source("B").dest("C").directed(true).build();
     private static final Edge EDGE_CA = new Edge.Builder().group(TestGroups.EDGE).source("C").dest("A").directed(true).build();
@@ -40,7 +40,7 @@ public class ExtractWalkEdgesTest {
     @Test
     public void shouldReturnEdgesFromWalkObject() {
         // Given
-        final Function<Walk, Iterable<Set<Edge>>> function = new ExtractWalkEdges();
+        final Function<Walk, Set<Entity>> function = new ExtractWalkEntitiesFromHop(2);
         final Walk walk = new Walk.Builder()
                 .edge(EDGE_AB)
                 .entity(ENTITY_B)
@@ -50,11 +50,9 @@ public class ExtractWalkEdgesTest {
                 .build();
 
         // When
-        final Iterable<Set<Edge>> results = function.apply(walk);
+        final Set<Entity> results = function.apply(walk);
 
         // Then
-        assertThat(results, containsInAnyOrder(Sets.newHashSet(EDGE_AB),
-                Sets.newHashSet(EDGE_BC),
-                Sets.newHashSet(EDGE_CA)));
+        assertThat(results, contains(ENTITY_C));
     }
 }
