+++ core/store/src/test/java/uk/gov/gchq/gaffer/store/operation/handler/MapHandlerTest.java
@@ -15,22 +15,16 @@
  */
 package uk.gov.gchq.gaffer.store.operation.handler;
 
-import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import org.junit.Before;
 import org.junit.Test;
 
-import uk.gov.gchq.gaffer.commonutil.CloseableUtil;
 import uk.gov.gchq.gaffer.commonutil.TestGroups;
 import uk.gov.gchq.gaffer.data.element.Edge;
 import uk.gov.gchq.gaffer.data.element.Entity;
 import uk.gov.gchq.gaffer.data.graph.Walk;
-import uk.gov.gchq.gaffer.data.graph.function.ExtractWalkEdges;
-import uk.gov.gchq.gaffer.data.graph.function.ExtractWalkEdgesFromHop;
-import uk.gov.gchq.gaffer.data.graph.function.IterableFunction;
-import uk.gov.gchq.gaffer.exception.SerialisationException;
-import uk.gov.gchq.gaffer.jsonserialisation.JSONSerialiser;
+import uk.gov.gchq.gaffer.data.graph.function.walk.ExtractWalkEdgesFromHop;
 import uk.gov.gchq.gaffer.operation.OperationChain;
 import uk.gov.gchq.gaffer.operation.OperationException;
 import uk.gov.gchq.gaffer.operation.impl.Map;
@@ -43,10 +37,13 @@ import uk.gov.gchq.gaffer.store.operation.OperationChainValidator;
 import uk.gov.gchq.gaffer.store.optimiser.OperationChainOptimiser;
 import uk.gov.gchq.gaffer.user.User;
 import uk.gov.gchq.koryphe.ValidationResult;
+import uk.gov.gchq.koryphe.impl.function.FirstItem;
+import uk.gov.gchq.koryphe.impl.function.IterableConcat;
+import uk.gov.gchq.koryphe.impl.function.IterableFunction;
+import uk.gov.gchq.koryphe.impl.function.NthItem;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 import java.util.function.Function;
@@ -373,25 +370,18 @@ public class MapHandlerTest {
         final Iterable<Walk> walks = Arrays.asList(walk, walk1);
 
         final Map<Iterable<Walk>, Iterable<Edge>> map = new Map.Builder<Iterable<Walk>>()
-//                .input(walks)
+                .input(walks)
                 .first(new IterableFunction.Builder<Walk>()
                         .first(new ExtractWalkEdgesFromHop(1))
                         .then(new FirstItem<>())
                         .build())
                 .build();
 
-        try {
-            System.out.println(new String(JSONSerialiser.serialise(map, true)));
-        } catch (SerialisationException e) {
-            e.printStackTrace();
-        }
-
         final ToVertices toVertices = new ToVertices.Builder()
                 .edgeVertices(ToVertices.EdgeVertices.SOURCE)
                 .build();
 
         final ToSet<Object> toSet = new ToSet<>();
-
 
         final OperationChain<Set<?>> opChain = new OperationChain.Builder()
                 .first(map)
@@ -414,61 +404,5 @@ public class MapHandlerTest {
 
         // Then
         assertThat(results, contains("B"));
-    }
-
-    // To be removed after Koryphe 1.1.0
-    private static class IterableConcat<I_ITEM> implements Function<Iterable<Iterable<I_ITEM>>, Iterable<I_ITEM>> {
-        @Override
-        public Iterable<I_ITEM> apply(final Iterable<Iterable<I_ITEM>> items) {
-            return Iterables.concat(items);
-        }
-    }
-
-    // To be removed after Koryphe 1.1.0
-    private static class FirstItem<T> implements Function<Iterable<T>, T> {
-        @Override
-        public T apply(final Iterable<T> input) {
-            if (null == input) {
-                throw new IllegalArgumentException("Input cannot be null");
-            }
-            try {
-                return Iterables.getFirst(input, null);
-            } finally {
-                CloseableUtil.close(input);
-            }
-        }
-    }
-
-    // To be removed after Koryphe 1.1.0
-    private static class NthItem<T> implements Function<Iterable<T>, T> {
-        public int getSelection() {
-            return selection;
-        }
-
-        public void setSelection(final int selection) {
-            this.selection = selection;
-        }
-
-        public NthItem() {
-            // Empty
-        }
-
-        public NthItem(final int selection) {
-            this.selection = selection;
-        }
-
-        private int selection;
-
-        @Override
-        public T apply(final Iterable<T> input) {
-            if (null == input) {
-                throw new IllegalArgumentException("Input cannot be null");
-            }
-            try {
-                return Iterables.get(input, selection);
-            } finally {
-                CloseableUtil.close(input);
-            }
-        }
     }
 }
